% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tingor.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Odds Ratio and confidence intervals}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{a numeric value for probability}

\item{se}{a numeric value for stand error}

\item{siglevel}{a numeric value for significant level}

\item{roundto}{a numeric valuefor decimal places}
}
\value{
a table for Odds Ratio, 95\% lower bound and upper bound
}
\description{
a function that returns the odd ratio result and 95\% confidence intervals
}
\examples{
coef <- c(0.2538319,0.2534135,0.5947923 )
se <- c(0.2538319, 0.2534135, 0.5947923 )
siglevel = 0.05
roundto = 4
OR_95CI(coef, se, siglevel, roundto)
}
\author{
Ting Lin
}
