% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{y}{binary value with 1 and 0}
\item{x1}{numeric value}
\item{x2}{binary value with 1 and 0}
}
}
\usage{
toydata
}
\description{
Data from our lecture :)
}
\keyword{datasets}
